+++ R/word.R
@@ -81,20 +81,45 @@ reassign_lines2 <- function(lines, cont, hunk.diff) {
 }
 ## Helper Function for Mapping Word Diffs to Lines
 ##
-## Used when we're doing a wrapped diff for atomic vectors.  We have been
-## through several iterations trying to get the most intuitive behavior and the
-## result is a fairly non-intuitive and likely inefficient algorithm.  It works
-## for the most part, so we leave it as is, but is long, messy, and could
-## probably be replaced by a far more elegant solution.
+## Used when we're doing a wrapped diff for atomic vectors.  We expect to
+## receive `tar/cur.dat` with any meta data lines (e.g. factor levels, time
+## series meta data) removed already.  The challenge is we need to then be able
+## to re-map back each word back to the line it was on originally before
+## unwrapping.  This may include adding padding line blanks in the case one hunk
+## displays across more lines than another.
+#
+## This function does two things: inserts padding lines when hunks in one object
+## end up longer than in the other and similar lines don't align, and computes
+## mock character strings that will be used by to force alignments. We
+## manufacture unique strings that either match or don't match across the two
+## objects depending on the word contents of each line, and then pass those
+## back as the `comp` component of the `tar.dat` and `cur.dat` returned.  The
+## subsequent line diff will use `comp` and cause the relevant lines to be lined
+## up.  This is inefficient and round-about, but has the huge benefit of
+## allowing us to plug in the wrapped diff into our existing line diff
+## infrastructure
 ##
-## @param tar.ends and cur.ends are the indices of the last elements in each line
-##   of the vector
-## @param tar.dat and cur.dat are the data
-## @param tar.ind and cur.ind seem to be used by augment, but don't totally
-##   remember what for.  They can have positive and negative values
+## Note that in "word" mode the returned values may be longer than the input ones
+## as it may be necessary to add lines to get things to match-up.  Added lines
+## are indicated by TRUE values in the `fill` component of the `*.dat` return
+## values
+##
+## We have been through several iterations trying to get the most intuitive
+## behavior and the result is a fairly non-intuitive and likely inefficient
+## algorithm.  It works for the most part, so we leave it as is, but is long,
+## messy, and should be replaced by a more elegant solution.
+##
+## @param tar.ends and cur.ends are the indices of the last elements in each
+## line of the vector
+## @param tar.dat and cur.dat are the data, see `line_diff` body for detailed
+##   description of them (about 100 lines in).  Note that the data has been
+##   subset to just the portion of it that has row headers (e.g. excluding
+##   factor levels, etc.)
+## @param tar/cur.ends the position in the unwrapped vector of the last "word"
+##   in each line.
 
 word_to_line_map <- function(
-  hunks, tar.dat, cur.dat, tar.ends, cur.ends, tar.ind, cur.ind
+  hunks, tar.dat, cur.dat, tar.ends, cur.ends
 ) {
   # Once we've done all the replication and disambiguation, we need to make sure
   # diff hunks have the same number of lines.  Start mix lines or start/end mix
@@ -158,9 +183,9 @@ word_to_line_map <- function(
   h.cont <- vapply(hunks, "[[", logical(1L), "context")
 
   # Compute what indices are in each lines; we are going to use this to
-  # categorize what time of line this is; some of this might be duplicative with
+  # categorize what type of line this is; some of this might be duplicative with
   # what we did earlier, but that was so long ago I don't want to get back into
-  # it
+  # it.
 
   tar.idx <- Map(seq, c(1L, head(tar.ends, -1L) + 1L), tar.ends, by=1L)
   cur.idx <- Map(seq, c(1L, head(cur.ends, -1L) + 1L), cur.ends, by=1L)
@@ -342,42 +367,27 @@ word_to_line_map <- function(
       else tar.lines.f2[[i]] <- short.pad
   } }
   # Augment the input vectors by the blanks we added; these blanks are
-  # represented by NAs in our index vector so should be easy to do
+  # represented by NAs in our index vector.
 
-  augment <- function(dat, lines, ind) {
+  augment <- function(dat, lines) {
     lines.u <- unlist(lines)
     lines.len <- length(lines.u)
     for(i in names(dat)) {
-      i.vec <- dat[[i]]
-      hd.ind <- seq_along(i.vec) < min(ind)
-      tl.ind <- seq_along(i.vec) > max(ind)
-      hd <- i.vec[hd.ind]
-      tl <- i.vec[tl.ind]
-      bod <- vector(typeof(i.vec), length(lines.u))
-      bod[!is.na(lines.u)] <- i.vec
+      i.vec <- vector(typeof(dat[[i]]), length(lines.u))
+      i.vec[!is.na(lines.u)] <- dat[[i]]
       if(i == "word.ind") {
-        bod[is.na(lines.u)] <- list(.word.diff.atom)
+        i.vec[is.na(lines.u)] <- list(.word.diff.atom)
       } else if (i == "fill") {
         # warning: this is also used/subverted for augmenting the original
         # indices so think before you change it
-        bod[is.na(lines.u)] <- TRUE
+        i.vec[is.na(lines.u)] <- TRUE
       }
-      dat[[i]] <- c(hd, bod, tl)
+      dat[[i]] <- i.vec
     }
     dat
   }
-  tar.dat <- augment(tar.dat, tar.lines.f2, tar.ind)
-  cur.dat <- augment(cur.dat, cur.lines.f2, cur.ind)
-
-  # Also need to augment the indices so we can re-insert properly; we subvert
-  # the fill logic since that will make sure
-
-  tar.ind.a <-
-    augment(list(fill=!logical(length(tar.ind))), tar.lines.f2, tar.ind)
-  tar.ind.a.l <- unname(unlist(tar.ind.a))
-  cur.ind.a <-
-    augment(list(fill=!logical(length(cur.ind))), cur.lines.f2, cur.ind)
-  cur.ind.a.l <- unname(unlist(cur.ind.a))
+  tar.dat.aug <- augment(tar.dat, tar.lines.f2)
+  cur.dat.aug <- augment(cur.dat, cur.lines.f2)
 
   # Generate the final vectors to do the diffs on; these should be unique
   # and matching for the matches, and unique and mismatching for the
@@ -395,8 +405,9 @@ word_to_line_map <- function(
   }
   neg.nums <- sum(!tar.match, !cur.match)
 
-  strings <-
-    make_unique_strings(pos.nums + neg.nums, c(tar.dat$raw, cur.dat$raw))
+  strings <- make_unique_strings(
+    pos.nums + neg.nums, c(tar.dat.aug$raw, cur.dat.aug$raw)
+  )
   strings.pos <- strings[seq.int(pos.nums)]
   strings.neg <- tail(strings, neg.nums)
   if(neg.nums + pos.nums != length(strings)) {
@@ -405,11 +416,11 @@ word_to_line_map <- function(
     # nocov end
   }
 
-  tar.dat$comp[tar.ind.a.l][tar.match] <- strings.pos
-  cur.dat$comp[cur.ind.a.l][cur.match] <- strings.pos
-  tar.dat$comp[tar.ind.a.l][!tar.match] <- head(strings.neg, sum(!tar.match))
-  cur.dat$comp[cur.ind.a.l][!cur.match] <- tail(strings.neg, sum(!cur.match))
-  list(tar.dat=tar.dat, cur.dat=cur.dat)
+  tar.dat.aug$comp[tar.match] <- strings.pos
+  cur.dat.aug$comp[cur.match] <- strings.pos
+  tar.dat.aug$comp[!tar.match] <- head(strings.neg, sum(!tar.match))
+  cur.dat.aug$comp[!cur.match] <- tail(strings.neg, sum(!cur.match))
+  list(tar.dat=tar.dat.aug, cur.dat=cur.dat.aug)
 }
 # Pull out mismatching words from the word regexec; helper functions
 
@@ -455,16 +466,10 @@ reg_apply <- function(reg, ends, mismatch) {
 # Modify `tar.dat` and `cur.dat` by generating `regmatches` indices for the
 # words that are different
 #
-# If `diff.mode` is "wrap", then line up lines based on the word matches and
-# mismatches contained there-in.  This is done by generating new strings
-# that match or don't depending on the word contents, and then passing those
-# back as the `comp` component of the `tar.dat` and `cur.dat` returned.  The
-# subsequent line diff will cause the relevant lines to be lined up.
-#
-# Note that in "word" mode the returned values may be longer than the input ones
-# as it may be necessary to add lines to get things to match-up.  Added lines
-# are indicated by TRUE values in the `fill` component of the `*.dat` return
-# values
+# If `diff.mode` is "wrap", then wrapped atomic vector output is unwrapped and
+# the diff is carried out in the unwrapped form, and then re-assembled.  See
+# `word_to_line_map` for details in how its done.  Return values may be longer
+# than input in this mode.
 #
 # `match.quotes` will make "words" starting and ending with quotes; it should
 # only be used with atomic character vectors or possibly deparsed objects.
@@ -535,14 +540,6 @@ diff_word2 <- function(
     tar.unsplit, cur.unsplit, etc=etc, diff.mode=diff.mode, warn=warn
   )
   # Need to figure out which elements match, and which ones do not
-  #
-  # questions about the `abs`; should it be applied to both `tar` and `cur`?
-  # can definitely have negative numbers in `x$A`; stuff seems to work fine, but
-  # it seems like this should cause problems.  Maybe this only ever runs in
-  # context mode so it's fine?  Odd part is that in browsing when debuggin at
-  # some point we most definitely saw negative numbers in x$A, although it is
-  # possible we were at the wrong spot in the call stack and looked at the
-  # original line diff hunks instead of the word diff ones...
 
   hunks.flat <- diffs$hunks
   tar.mism <- unlist(
@@ -563,22 +560,54 @@ diff_word2 <- function(
   cur.dat$word.ind[cur.ind] <- reg_apply(cur.reg, cur.ends, cur.mism)
 
   # If in wrap mode (which is really atomic mode), generate a spoofed
-  # `comp` vector that will force the line diff to align in a way that respects
-  # the word differences.  This is inefficient and round-about, but has the
-  # huge benefit of allowing us to plug in the wrapped diff into our existing
-  # line diff infrastructure
+  # `comp` vector (see word_to_line_map)
   #
   # Note that we're only operating on a subset of the data via tar.ind and
-  # cur.ind
+  # cur.ind, these are supposed to be the contiguous block of lines that have
+  # row headers.
 
+  tar.dat.fin <- tar.dat
+  cur.dat.fin <- cur.dat
   if(diff.mode == "wrap") {
+    tar.dat.ind <- lapply(tar.dat, '[', tar.ind)
+    cur.dat.ind <- lapply(cur.dat, '[', cur.ind)
     word.line.mapped <- word_to_line_map(
-      hunks.flat, tar.dat, cur.dat, tar.ends, cur.ends, tar.ind, cur.ind
+      hunks.flat, tar.dat.ind, cur.dat.ind, tar.ends, cur.ends
     )
-    tar.dat <- word.line.mapped$tar.dat
-    cur.dat <- word.line.mapped$cur.dat
+    # Merge back the mapped data, need to account for possiblity of padding
+    # lines being added.
+
+    tar.len.old <- length(tar.dat[[1L]])
+    cur.len.old <- length(cur.dat[[1L]])
+
+    tar.ind.lo <- seq_len(head(tar.ind, 1L) - 1L)
+    tar.ind.hi <- seq_len(tar.len.old - tail(tar.ind, 1L)) + tail(tar.ind, 1L)
+    cur.ind.lo <- seq_len(head(cur.ind, 1L) - 1L)
+    cur.ind.hi <- seq_len(cur.len.old - tail(cur.ind, 1L)) + tail(cur.ind, 1L)
+
+    interleave <- function(idx, new, old, lo, hi)
+      c(old[[idx]][lo], new[[idx]], old[[idx]][hi])
+
+    tar.dat.fin <- setNames(
+      lapply(
+        seq_along(tar.dat), interleave,
+        new=word.line.mapped[['tar.dat']], old=tar.dat,
+        lo=tar.ind.lo, hi=tar.ind.hi
+      ),
+      names(tar.dat)
+    )
+    cur.dat.fin <- setNames(
+      lapply(
+        seq_along(cur.dat), interleave,
+        new=word.line.mapped[['cur.dat']], old=cur.dat,
+        lo=cur.ind.lo, hi=cur.ind.hi
+      ),
+      names(cur.dat)
+    )
   }
-  list(tar.dat=tar.dat, cur.dat=cur.dat, hit.diffs.max=diffs$hit.diffs.max)
+  list(
+    tar.dat=tar.dat.fin, cur.dat=cur.dat.fin, hit.diffs.max=diffs$hit.diffs.max
+  )
 }
 # Make unique strings
 #