+++ tests/testthat/testthat.diffChr.R
@@ -154,3 +154,15 @@ test_that("Nested dots issue 134, h/t Noam Ross", {
     )
   )
 })
+test_that("Newlines in input, issue 135, h/t Flying Sheep", {
+  a <-     'A Time Series:\n[1] 1 2 3 4'
+  b <-     'A Time Series:\n[1] 9 4 1 4'
+  expect_equal(
+    c(as.character(diffobj::diffChr(a, b, format = 'raw'))),
+    c("< a               > b             ",
+      "@@ 1,2 @@         @@ 1,2 @@       ", 
+      "  A Time Series:    A Time Series:", 
+      "< [1] 1 2 3 4     > [1] 9 4 1 4   ")
+  )
+})
+