+++ NEWS.md
@@ -2,6 +2,14 @@
 
 ## v0.2.3
 
+This is a bugfix release.
+
+* [#136](https://github.com/brodieG/diffobj/issues/136): Documentation for
+  `ignore.white.space` (h/t @flying-sheep) and `max.diffs` parameters listed
+  incorrect defaults.
+* [#135](https://github.com/brodieG/diffobj/issues/135): Incorrect handling of
+  potential meta data strings when unwrapping atomics would cause a "wrong sign
+  in by argument" error (h/t @flying-sheep).
 * [#134](https://github.com/brodieG/diffobj/issues/134): Forwarding `...` to
   `diff*` functions no longer breaks substitution of arguments for diff banners
   (h/t @noamross)..