+++ NEWS.md
@@ -1,5 +1,24 @@
 # diffobj
 
+## v0.2.3
+
+This is a bugfix release.
+
+* [#136](https://github.com/brodieG/diffobj/issues/136): Documentation for
+  `ignore.white.space` (h/t @flying-sheep) and `max.diffs` parameters listed
+  incorrect defaults.
+* [#135](https://github.com/brodieG/diffobj/issues/135): Incorrect handling of
+  potential meta data strings when unwrapping atomics would cause a "wrong sign
+  in by argument" error (h/t @flying-sheep).
+* [#134](https://github.com/brodieG/diffobj/issues/134): Forwarding `...` to
+  `diff*` functions no longer breaks substitution of arguments for diff banners
+  (h/t @noamross)..
+* [#133](https://github.com/brodieG/diffobj/issues/133): `diffFile` considers
+  files with equal content but different locations to be `all.equal` now (h/t
+  @noamross).
+* [#132](https://github.com/brodieG/diffobj/issues/132): Duplicate pager slot
+  for baseline `Pager` removed (h/t Bill Dunlap).
+
 ## v0.2.2
 
 * Set `RNGversion()` due to changes to sampling mechanism.